Adding "Search" preamble to label for search input
Prior to this change, when I would enter the `input[name=q]`` field in a
screen reader the field was announced as "CurateND, edit text" (or
"OneSearch"). By adding the "Search" preamble (but hiding it on the
screen), now when I enter the `input[name=q]`, the screen reader
announces "Search CurateND, edit text". I believe this provides a
greater context.

Note: When I go to Google's main page and enter the big input box the
label says "Search"

Related resource:

* http://juicystudio.com/article/screen-readers-display-none.php#providecontext