Improving search application accessibility
The document has a complicated visual UI:

* A region that indicates the search option to use; That region can be
  clicked to change the option.
* Based on the selected search option, the label of the `input[name=q]`
  should vary.

This breaks apart the labels into their two functional parts:

* Label the selecting mechanism
* Label the input mechanism

As I was exploring using auto-focus, I found that when the
`input[name=q]` receives focus, the screen reader does not read the
label (and thus doesn't clue in the person that there alternate search
options).

Note: The label tag remains as it is a visual indicator
(e.g. the carrot).