Add logout button to personal and courses page.
Add a logout button to personal and courses pages to facilitate
accessibility of button that is normally located in the dropdown ‘my
account’ menu.
When users tab through the menu they go directly to the personal page
upon pressing enter to be in alignment with the functionality of the
other dropdown menus. Switching between personal and courses pages was
possible through tabbing the interface, but logging out was not
possible. So a button was added.